+++ kanso/framework/exception/ErrorLogger.php
@@ -99,6 +99,7 @@ class ErrorLogger
         'DATE    : '.date('l jS \of F Y h:i:s A', time())."\n".
         'TYPE    : '.$this->errType().' ['.$this->exception->getCode()."]\n".
         'URL     : '.$this->environment->REQUEST_URL."\n".
+        'REFERER : '.$this->environment->REFERER."\n".
         'CLASS   : '.$this->errClass()."\n".
         'FILE    : '.$this->exception->getFile()."\n".
         'LINE    : '.$this->exception->getLine()."\n".